Silence most of the warnings in the x86_64 build, and kill some Real Bugs in the process.  (sanity checked in an x86 uml build, as well)
- include linux/init.h from some modules that use __init and company
- x86_64 has 64 bit longs but uses 'long long' for u64 in the kernel.  we
  add its own specific section to LP*64 definitions.  Sadly, there are still
  a huge number of warnings from this gcc when we try to LP*64 print uint64_t
  obdo arguments.
- use min_t in some defines to avoid the duplicate const prefacing that comes
  with nesting min()
- LPX64 printing -1UL doesn't work so well
- i_nlink can be any of unsigned short, int, or long.  when printing cast
  to unsigned long.
- kern_addr_valid takes a 'unsigned long' argument and shifts it.. sigh.
- don't #define EXPORT_SYMTAB in the source, that's the job of the kernel
  makefiles
- x86_64/suse has THREAD_SIZE defined in userspace
- s/long flags/unsigned long flags/.  Again.
- va_list ap = NULL; is not legal, as far as I know.  it doesn't seem
  to actually do anything in this code path either.
- x86_64 passes CONFIG_X86 but had thread->rsp as the stack pointer.  this
  code is going away eventually anyway.
- print a pointer with %p, not %x
- some portals callback pointer definitions used size_t while some nal's
  functions used ptl_size_t in the function defintions.  I switched the
  functions over to avoid pushing changes on all the nals.

+++ lustre/llite/llite_nfs.c
@@ -97,9 +97,9 @@ static struct dentry *ll_iget_for_nfs(struct super_block *sb, unsigned long ino,
             || (generation && inode->i_generation != generation)
             ){
                 /* we didn't find the right inode.. */
-              CERROR(" Inode %lu, Bad count: %d %d or version  %u %u\n",
+              CERROR(" Inode %lu, Bad count: %lu %d or version  %u %u\n",
                         inode->i_ino, 
-                        inode->i_nlink, 
+                        (unsigned long)inode->i_nlink, 
                         atomic_read(&inode->i_count), 
                         inode->i_generation, 
                         generation);
